MB-64484 Use quotes for identifiers in optimizer hints
Currently the optimizer hints in explain just use the identifiers
without any quotes, e.g. keyspace name, index name, etc. The issue
with this bug is that if an identifier requires escape (quoted
within backticks), such as `travel-sample`, such identifiers are
not properly quoted in the optimizer hints specifications, and as
such will not be properly recognized if needed to be parsed (e.g.
when shipped as prepared plan to a remote node).

The fix is to add proper quotes for all identifiers in optimizer
hints. The USE_HASH hint requires special handling since it has
the optional "/BUILD" or "/PROBE" option.

Change-Id: I8bfcf04315d8a0b9b05513de5af8a72635ac79ae
Reviewed-on: https://review.couchbase.org/c/query/+/220420
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker