MB-64596 Combine and simplify spans for IntersectSpans ...    ... when multiple array predicate on the same array key exists
This is an extension of the fix for MB-61549.

When we have an array index with FLATTEN_KEYS, and multiple array
predicate exists on the same array index key (multiple sub-predicate
within each array predicate due to FLATTEN_KEYS), we need to make
sure the index spans generated are from the same array predicate,
i.e. do not generate (combining spans) from different array
predicates. An arrayId is generated for this purpose. When combining
spans from different index keys, if the arrayId do not match, we
instead generate a _WHOLE_SPANS for the 2nd index key.

The issue with this bug is that due to the above fix, if both of
the array predicates have sub-predicate on different array keys
(within the FLATTEN_KEYS expression), then we end up generating
4 different index scans under the IntersectScan, 2 of them have
proper index spans but the other 2 have only valid index span
for the 1st array index key, and _WHOLE_SPANS for the 2nd array
index key. This will not cause wrong result, just inefficient
execution (the 2 index scans with _WHOLE_SPANS for the 2nd
array index key is doing unnecessary work).

To fix this, we need to detect situations where an arrayId is
used to generate spans under an IntersectSpan (resulting from
multiple array predicates), and add an extra step to try to
simplify the spans generated. The extra step is to see for the
same arrayId used, if we have previously generated an span
already, make sure we "constrain" the spans such that we take
the "minimum" of all spans using the same arrayId. This should
modify the _WHOLE_SPANS generated to be the more appropriate
span, and the (existing) dedup step later should then get rid
of the redundant span.

Previously the Constrain() function for SargSpans only deal
with the first range since it only has a single range. For
this purpose we need to extend that to cover multiple ranges.
The Constrain() and ConstrainTerm() functions are modified
to deal with multiple ranges under an index span.

Change-Id: Id20afc3cb4ac6790f16271c0b47b9197ba8654f3
Reviewed-on: https://review.couchbase.org/c/query/+/220907
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>