MB-64613 Set Onclause to be pushable when transforming ...    ... an ANSI outer join to an inner join
The query engine performs ANSI outer join to inner join
transformation when appropriate. For an inner join, the ON
clause of the join is considered "pushable" in most cases,
i.e. (part of) the ON clause can be evaluated before the join
itself, or be pushed to under the join. However there are
rare situations where the ON clause cannot be pushed, e.g.
if the ON clause references an outer table (which appears
before this join). The query engine performs appropriate
marking of the ON clause (pushable or not) for an inner
join. The issue with this bug is that when a join is
originally specified as as outer join, but gets transformed
into an inner join internally, we are not marking the ON
clause as "pushable" in this case. This results in the ON
clause being added as join filter later on again, thus
duplicating the filter, and triggers a sanity check during
the join enumeration process.

The fix is to apply the same logic as regular inner joins,
in marking the ON clause as pushable or not, if an outer
join is transformed into an inner join.

Change-Id: Ia57440b0a83cece32935621cc2e42a2286fd7b21
Reviewed-on: https://review.couchbase.org/c/query/+/220937
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker