MB-64666 Check for duplicated probe expression references ...    ... when performing join filter for hash join
For hash join the optimizer does cost-based decision for whether
to construct a bloom-filter from the build side and push it to
the probe side index scan to reduce fetch size. With join
enumeration on, since we pre-construct potential bloom filter
early on, it is possible that we encounter the same bloom filter
information multiple times for the same index. Current logic
checks whether the existing bloom filter information is the same
as the new one being constructed, and allow it if it's the same.
The issue with this particular bug is that we have a 3-way join
on the same column, and thus the same build expression can be
used to build a bloom filter for 2 separate receivers, and
current logic always just put the bloom filter expressions in
an array, and thus we could be putting the same column twice,
which later on may result in error in determining whether the
bloom filter information is the same or not, triggering a
sanity check failure.

The fix is to first check whether the probe expression is
already in the array, before appending it to the array.

Change-Id: I6cf4d1c35c86dca3e7b830297e6bdd8e992c305b
Reviewed-on: https://review.couchbase.org/c/query/+/221119
Well-Formed: Restriction Checker
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>