MB-64684 Remove binding variable from correlation map ...    ... when binding is popped (go out of scope)
This bug is somewhat similar to MB-64408, except in this bug
the correlation reference is a binding variable. When we have
nested bindings, if the inner level references the outer-level
binding variable the variable name is added to the correlation
map of the formalizer. Later when (eventually) the outer level
binding is popped (i.e. going out of scope), the binding
variable still remains in the correlation map, resulting in a
sanity error.

The fix is to check and remove a binding variable from the
correlation map when the binding variable is popped.

Change-Id: I0674b22fa02798f35de9ff8251b83291da400db6
Reviewed-on: https://review.couchbase.org/c/query/+/221179
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>