Add diagnostic code to capture the last results sent when a request is has TRACE log level set.
This comes about as a result of a case where it is claimed the server intermittently sends the incorrect results to a client and allows _some_ record of what the server actually sent.

No difference in Q1 TCO figures with this code present (but not in use - i.e. without setting LOGLEVEL).

Change-Id: I4216ebc053c58741f8c12b32b148c96b9e2f71c8
Reviewed-on: https://review.couchbase.org/c/query/+/221229
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Donald Haggart <donald.haggart@couchbase.com>