MB-64722 Allow static variables as argument to _timeseries function
This bug is a regression from MB-64442.
We introduced a new function to check "static" references that excludes
static variables (WITH alias, function arguments), and use that new
function in places where we do not want to include static variables.

For the second argument of _timeseries function we should allow static
variable references.

The fix is to undo a couple of changes from the previous fix.

Change-Id: I29f956a9f49dcc56168165713b4083bcfeafbeba
Reviewed-on: https://review.couchbase.org/c/query/+/221405
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>