MB-64752 Do not add redundant filters
The problematic statement has redundant filters (same filter specified
multiple times) involving query parameters. This results in multiple
index spans being generated and the index spans are marked as "not
exact", which in turn disables index pushdown.

The fix is to add logic to detect redundant filters being used when
adding filters to baseKeyspace, such that we only use 1 copy of the
filter.

Change-Id: I026680a5a22fd743abe5077a7367b0a032e36416
Reviewed-on: https://review.couchbase.org/c/query/+/221493
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/221536