MB-64917 Keep track of outer unnests on primary keyspace
This bug is very similar to MB-63069 except it uses OUTER UNNEST
instead of (INNER) UNNEST. The fix for MB-63069 did not cover
this case since currently we only keep track of dependent inner
unnests for the primary keyspace (for unnest scan considerations).

Add logic to also keep track of dependent outer unnests as well
for the primary keyspace. This information will be used when
checking whether the primary keyspace contains ANY unnests.

Change-Id: Id31c04ed45c427204f4d6cc981cc705498c31103
Reviewed-on: https://review.couchbase.org/c/query/+/222157
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>