MB-65055 Properly handle WITH alias in plans generated with older server
As part of the change for the new recursive-CTE support in 7.6, the way
we marshal/unmarshal a WITH clause changed, as a result, in case of a
prepared statement, if a rolling upgrade is done, then the prepared
statement is propagated from the older version (e.g. 7.2) query node to
a newer version (e.g. 7.6) query node. If a WITH clause is used in the
prepared statement, due to the different ways of marshal/unmarshal of
a WITH clause, we end up with a missing WITH alias in the newer server,
resulting in incorrect usage and wrong result from such "migrated"
prepared statements.

The fix is to check and detect situations where the plan might be from
an older version, and get the WITH alias from a different variable
saved in the plan.

The unmarshal of binding variables are also out-of-date, missing a
couple of newer fields. This is also fixed.

Change-Id: I265f7722643e5e9a19dcb0fe88744e79f0b787f2
Reviewed-on: https://review.couchbase.org/c/query/+/222567
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/223687
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>