MB-65085: Add query support for "Hybrid" client cert auth setting...
Client certificate authentication has a new value "Hybrid".

* Internal Client Certificate authentication is required for
inter-service encrypted communication when:
    i.  Now, When N2N encryption is set to "All" or "Strict"
    AND
    ii. Client Certificate Authentication set to "Mandatory" or
    "Hybrid"

When Client Certificate Authentication setting is set to "Hybrid":

i. For internal clients:
    a.  If n2n encyption is enabled, all encrypted communication must
    authenticate using internal client certificate.

    b. if n2n is disabled, the internal client will use Basic Auth to
    authenticate.

ii. For external clients:

The behaviour must be the same as when the Client Cert Auth setting is
"Enabled".  i.e a request sent to the encrypted port can authenticate
with Basic Auth or Client Certificate.

Changes in the PR:
* cbauth.TLSConfig has a new field `ShouldClientsUseClientCert`. It is
set to true when Client Certificate Authentication setting is
"Mandatory" or "Hybrid".

* Now when n2n encryption is enabled, Query will check
TLSConfig.ShouldClientsUseClientCert and determine whether to
authenticate using internal client certificate or not.

Change-Id: If61012b2955e8a6eda899221574e10e0e4a4574a
Reviewed-on: https://review.couchbase.org/c/query/+/223069
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>