MB-65176 Improve parsing of full path and query context
The issue with this bug is failure of UPDATE STATISICS command
to generate optimizer statistics, which then results in
sub-optimal query plans for join queries.

The 4 buckets used in this test all have single-character names:
R, S, T and U. As part of the fix for MB-40444 we modified the
way a full path name is parsed, in order to fix an issue in
earlier code lines (see MB-65225). However, that additional fix
now results in this issue where we did not properly recognize
the bucket name in a full path. Note the issue that MB-65225
is the use of a query context without the namespace, e.g.
b.s, while the issue with this bug is a full path with namespace,
but has a single-character bucket name, e.g.
default:R._default._default.

To fix this, we modify the way we check and adjust for the
situation when backticks are used around identifiers, instead of
checking for array position, which could sometimes fail with a
single-character identifier, we now explicitly set a boolean
indicator, specifically for the case when backticks are used,
such that for the character immediately after the closing of a
backtick, we can properly detect the closure of a backtick and
set the ending of an identifier properly.

Change-Id: I12cbdc74c7e536fdc89be8a98d18e956887d3464
Reviewed-on: https://review.couchbase.org/c/query/+/223141
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>