MB-65225 Improve parsing of full path and query context
The issue with this bug is that when parsing a query context
or a full path, we currently detect the end of a backtick by
checking the array position, which does not always work
properly if we start the query context with a bucket that has
a single-character name. In this case the current code does
not properly differentiate between the previous character
being a (closing) backtick and a single-character identifier.

To fix this, we modify the way we check and adjust for the
situation when backticks are used around identifiers, instead of
checking for array position, which could sometimes fail with a
single-character identifier, we now explicitly set a boolean
indicator, specifically for the case when backticks are used,
such that for the character immediately after the closing of a
backtick, we can properly detect the closure of a backtick and
set the ending of an identifier properly.

Change-Id: I77d1626e0a235863280ec9e6f8e0ebf636ab368b
Reviewed-on: https://review.couchbase.org/c/query/+/223142
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker