MB-65296: Fix FFDC captures not being collected in certain conditions
FFDC can be triggered by different events like memory exceeded,
queue full, etc.

Each event can trigger FFDC twice, but after the second capture,
the event must be "reset" before FFDC can be collected again.

The system determines that a capture is unnecessary since the
condition has remained unchanged for a long time.

The issue is that, if the system does not stop experiencing
the event, it will not be reset. And FFDC is not captured for
extended periods of time.

This PR aims to mitigate this by triggering a reset on every
event every 15 minutes.

Change-Id: Iaf26b20a97b0dde7fa2741dcd681ebbb21ed4274
Reviewed-on: https://review.couchbase.org/c/query/+/223455
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>