MB-60680 Avoid modifying expressions from indexer
The include expressions that we get from the indexer comes as a
slice of expressions, and current code modifies the elements
of the slice as we process them (formalization). This results in
the next call to get the include expressions to have formalized
expressions.

We need to make a copy of the slice such that we don't end up
modifying the original slice returned from the indexer.

This is not an issue for the index keys since we already make
a new slice for the keys.

Change-Id: I0cc45618ec43a82a25cc3af21ddac8eb8061ace3
Reviewed-on: https://review.couchbase.org/c/query/+/223508
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>