Unit test update
The unit test indexcbo starts to fail for me after the latest
changes for include columns. The reason being for whatever
reason the newly created indexes are not memory resident and
thus UPDATE STATISTICS did not properly gather index statistics
for the new indexes, resulting in failure in CBO costing.

Add a few queries to use the newly created indexes such that
the indexes are at least partially in memory before running
UPDATE STATISTICS.

Change-Id: I53743bee6a7ec407a351cd78de68e0b279f4e454
Reviewed-on: https://review.couchbase.org/c/query/+/223517
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>