MB-65203 Introduce StringValuePoolForIndex
When filtering on include columns, index service observed that
a lot of CPU is being spent on allocating and garbage collecting
"stringValue" objects during type conversion. To avoid such
allocations, a new "StringValuePoolForIndex" object is added.

It is indexed because index service knows the position of the
object getting decoded and it will always use the same wrapper
for that specific position. This ensures that the object is not
shared with other decoded values

Currently, only use case is include column filtering at index serivce.
This pools is not used else where.

Change-Id: Icab567fe08ef1bc895109397c2ac46b0b6d9c731
Reviewed-on: https://review.couchbase.org/c/query/+/223556
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>