MB-60680 Generate index spans for include columns
For index keys the planner currently generates index spans,
which is what the indexer uses for the actual index scan. For
include columns, currently we only pass the index filter
expressions to the indexer. For performance reasons it is
helpful to also generate index spans for the simple filters
on the include columns, such that the indexer can potentially
perform certain operations without needing to do the full
expression evaluation.

For CBO purpose it is also useful to have index spans for
the include columns as well.

We now generates index spans for include columns the same
way we generate index spans for index keys. The generated
index spans are more or less the same as regular index keys,
the only difference being that we only take TermSpans at
the end for include columns, no multi-spans (intersect or
union). If an intersect or union span is generated, we
simply discard it.

The generated index spans for include columns are only for
explain in this checkin. We currently do not pass that info
to the indexer. A change in the index6 interface is required
to pass this new information to the indexer.

There are also aspects of the index spans that may require
additional work:

 - the limit on number of spans generated
 - the ability to generate index spans for "non-leading"
   include columns

In addition, the generation of index spans does not change
the usage of index filters, i.e. the filter expressions for
the include columns are still passed to the indexer as they
are today. Whether this can be optimized later needs
further investigation.

Change-Id: I2ed960194227460c7e37472732f347b70f7c0a83
Reviewed-on: https://review.couchbase.org/c/query/+/223635
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>