MB-65357 Better handling of LIKE predicate with query parameters
When a LIKE predicate involves a query parameter, CBO attempts
to use a default selectivity when it detects that optimizer
statistics exists on the field expression in the LIKE predicate.
The issue was that we are using a formalized field expression
when the histogram key is a unformalized expression, resulting
in a mismatch.

The fix is to first remove the keyspace reference from the
histogram key before checking for histogram information.

Also fixed an issue where the LIKE expression did not have
a proper flag when selectivity information is being checked.

Change-Id: I8f7b0bd7024234821496a3c9976c9a2279760a36
Reviewed-on: https://review.couchbase.org/c/query/+/223679
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>