MB-65467 Proper memory accounting for filtered documents in UNNEST
The memory accounting for UNNEST is somewhat incomplete, in that we track
all the documents generated in UNNEST, however due to filter pushdown we
could discard certain documents if filter evaluation fails. Currently we
do not release the memory for discarded documents properly, and in case
of this bug, there is a large number of documents being discarded, resulting
in high memory usage (higher than actual usage).

The memory tracking for UNNEST was incomplete since memory quota was
intially implemented, and was improved in 7.6 by MB-52944 (which was
backported to 7.2 as MB-52959). However the symptom of this bug does not
show up in 7.2 since the tracking of size of annotated values were also
improved in 7.6 (MB-54614) which was not in 7.2. As a result, the memory
calculations in 7.2 was a vastly underestimation, and did not trigger
the memory quota error, while in 7.6 the memory calculation was much
more realistic. The side effect though was that since memory was not
properly released when documents are filtered out, we end up with a much
higher memory accounting in 7.6 which triggers the memory quota error.

The fix is to add proper memory release in case a document fails filter
evaluation.

Previous accounting for the "last" element of the array was done in a
special case (see MB-55688), which seems a bit confusing and relies on
special knowledge on how size calculation is done. We "revert" to the
1st approach mentioned in the review comment of MB-55688 and recalculates
the amount of memory that should be tracked for the last element
(this is necessary since the original input document is already tracked,
and we reuse the origin input document for the last element of the array).
This also makes it easier to handle the "release" part of the memory
tracking.

Change-Id: Iaab5309f24c32be629dae5cb07ac84461311eb60
Reviewed-on: https://review.couchbase.org/c/query/+/223939
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>