MB-65467 Improve memory tracking for Merge
Similar memory tracking issue exists in Merge as well. Specifically,
if any of the MERGE action (UPDATE/DELETE/INSERT) has a WHERE clause,
and the evaluation of the WHERE clause fails, we currently discard the
input item without properly releasing the tracked memory, which can
result in over-counting of memory usage.

In addition, for DELETE specifically, we may generate a much smaller
AnnotatedValue to send to the Delete operator (not the original item),
in which case we also should adjust the tracked memory.

Change-Id: Id91f7c4ef375d608f383162e70a65e52c50d68be
Reviewed-on: https://review.couchbase.org/c/query/+/224014
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker