MB-65474: Revise scheduling mechanism
Previously, if an AUS task from an old schedule had run on a given day,
a schedule change would not allow a new task from the new schedule to
run that same day unless the new schedule's start time was after the
old schedule's end time, even if the old task had already finished
execution. This was inconvenient.

This revision allows scheduling of AUS tasks from the new schedule
on the same day, as long as the task from the old schedule has
completed.

Example:
If a task from the old schedule runs from 1:00-2:00 and finishes
at 1:15, a new task from the new schedule can be scheduled
for 1:30-2:30. This was previously not allowed.

Details:
* If the task from the old schedule is still running, a new task from
the new schedule can only be scheduled if its start time is after the
old task’s end time, to prevent concurrent AUS tasks.
* If the task from the old schedule has finished running, a new task
from the new schedule can be scheduled on the same day, even if it
overlaps with the old schedule’s window.
* Add schedule version to the coordination document.

MB-65409: Modify error message

Change-Id: I18ede7628003df026c6692a84f817843bf5a49ee
Reviewed-on: https://review.couchbase.org/c/query/+/224260
Tested-by: Dhanya Gowrish <dhanya.gowrish@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>