MB-65551 Consider LIMIT when CBO considers intersect scan
When CBO considers whether to use an intersect scan with multiple
indexes, it uses costs for performing the entire index scan(s),
which makes sense since if an intersect scan is chosen by CBO then
the underlying index scans need to finish. However, there is a
special case where if a single index is chosen, then it is possible
to have the Limit operator reaching the necessary limit and stop
the plan earlier, before the index scan completes. This special
case only applies to when a single index scan is used. Currently
the CBO costing when considering intersect scans does not take
into consideration this special case, and thus can favor an
intersect scan over a single index scan.

The fix is to add handling of this special case for when LIMIT is
specified and a single index is used.

A separate issue identified is when a query parameter is used
in an equality predicate, if we have distribution information on
the field, we could get a much better estimate by using the
number of distinct values for the field, instead of just using
a default selectivity. This only applies to equality predicates.

Change-Id: I9eacdf0e4933f3c18f7dda8d7078aefab4c841b5
Reviewed-on: https://review.couchbase.org/c/query/+/224377
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>