MB-65002 Add retry mechanism for...
   ...natural requests failing when the generated statement
can't be parsed successfully.

We retry to see if we can get a response that corrects the
original syntax error. Retry is done again via the completions
api with a retry prompt.

Time spent retrying is accounted for under the nlretry phase.

Additionally, replace intermedite bytes.buffer with
json.Decoder to directly decode the response.

Change-Id: Ie6046e3fdd1658fc5d6e84289c36b475e0cb6b38
Reviewed-on: https://review.couchbase.org/c/query/+/224648
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Gaurav J <gaurav.jayaraj@couchbase.com>