MB-65746 Do not keep index with WHERE clause unnecessarily
This bug is a regression from the fix for MB-61014.
We have two indexes defined, one with more sargable keys and the
other with a WHERE clause. The issue with the previous bug
MB-61014 was that we were throwing away the index with WHERE
clause too early, and that index can actually be used for covering
purpose. However for this bug, covering is not possible, but we
end up with both indexes due to the previous fix, resulting in an
intersect scan unnecessarily.

The fix is to limit the previous fix to only covering cases.

Change-Id: I0635bbd605fb464869d9722997bf63a54922c4b2
Reviewed-on: https://review.couchbase.org/c/query/+/224818
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker