MB-65722: Fix AWR not using latest hostname in document key generation
A single node cluster without an explicit node name set, will have a
node name of "127.0.0.1". This will be used in the AWR document key
generation.

But when an exlicit node name is set, or a node is added to the
cluster, the host name of the node will change.

But AWR reporting continued to use the old node name, until the
AWR routine was restarted.

Fix:

whoAmI() is called every 20 minutes to get the latest node name.

When the node name is 127.0.0.1, a call to whoAmI() can cause a
pool refresh. Hence call this method every 20 minutes.

Change-Id: I10ac853e3d28eda541eb06f265c32bde655d866e
Reviewed-on: https://review.couchbase.org/c/query/+/224825
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Dhanya Gowrish <dhanya.gowrish@couchbase.com>