MB-65862. Recycle only when two values FALSE
As part of MB-63420 the value is recursively tracked and recycled.
It used this condition. Problem with if v or pv contains NULL function
retruns value.NULL_VALUE which is true.

         if v.Equals(pv) != value.TRUE_VALUE  {
         }

If there is no parallelism AggregateFinal input/output same which causes
early recycle.

MB-65246. Indirectly fixed this (ARRAY_AGG) we skipped. Change the
condition reflect right thing.

NOTE: AggregateIntial left as is because Track/Recycle done in
conjection so not an issue (It safe to track than not)

Change-Id: Ic3e30a3341c86381b71ce52b9e9cfc94338ecb8e
Reviewed-on: https://review.couchbase.org/c/query/+/225089
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Well-Formed: Restriction Checker