MB-65832 Proper handling of index group/aggregate pushdown ...    ... during join enumeration
This bug is a continuation of the fixes for MB-63702/MB-63947.
When a query contains a SubqueryTerm in a join, and the subquery
can use a covering index scan, since the SubqueryTerm is planned
multiple time during join enumeration, we delay covering
transformation for the SubqueryTerm until a final plan is chosen
and perform cover transformation at the end of join enumeration.

The issue with this bug is that if index group/aggregate pushdown
is done in the SubqueryTerm, then there is an extra step to also
perform cover transformation for the newly added index group and
index aggregates. This extra step is not currently done when
covering transformation is performed at the end of join
enumeration, resulting in improper representation of certain
expressions in the plan.

The fix is to also perform this extra step when covering
transformation is done at the end of join enumeration. To make
this work:

 - add a list of aggregates to each coveringSubqInfo entry
 - when a new coveringSubqInfo entry is generated, include the
   list of current aggregates from the builder (current Subselect)
 - separate the current logic for performing the extra step for
   covering transformation for index group/aggregate pushdown
   into two separate steps:
       - a "prep" step that performs expression transformation
         for the group/aggregate expressions, adding new "cover"
         expressions for group/aggregate, and generating several
         new "coverers", one for group, one for partial aggregate,
         one for full aggregate.
      - using the newly generated coverers to cover the rest of
        the expressions
 - when performing covering transformation at the end of join
   enumeration, do the same "pre" step as above, then use the
   newly generated coverers to perform cover transformation
   for relevant expressions in each operator of the final plan.

Change-Id: I5f832672aade546cfc202185cb5e6fc74fa27a00
Reviewed-on: https://review.couchbase.org/c/query/+/225122
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>