MB-65949  CBO may pick nested-loop join over hash join when ...    ... one side of the join is an expression or a subquery
When a join is being done and one side of the join is an expression
or a subquery, while the other side is a collection, since it is
very difficult to try to estimate a join selectivity between a
collection and an expression/subquery, we currently take the number
of distinct values for the join column in the collection and just
takes (1/distincts) as an estimate for join selectivity.

The issue in this bug is that the collection has a couple of filters
that selects very few rows from the collection, i.e. the selectiviy
for the table filtes are low. In this case using (1/distincts) is
a vast underestimation of the join selectivity, resulting in a
nested-loop join with the collection as the inner, which is
inefficient since the subquery produced a larger number of output.

While it is almost impossible to accurately estimate a proper
join selectivity in this case, we could improve on the previous
implementation by taking into consideration of the table filters,
i.e. try to estimate the number of distinct values for the join
column after the table filters have been applied.

To make this work, we need to know that a join selectivity estimation
falls in this category, i.e. was estimated from (1/distincts). The
current selectivity estimation code returns (in addition to selectivity)
only a boolean indicating whether DEFAULT selectivity was used. We now
use a flags instead of a boolean such that multiple flags can be passed.

The fix includes:

 - modification of selectivity calculation to return flags instead of
   boolean (this accounts for majority of the changes)
 - mark join selectivity with a new flag (SELEC_AVG_DISTINCT) when it
   is derived from (1/distincts). This is currently only done when one
   side of the join is an expression or a subquery term.
 - add a couple of new Filter flags to mark a join filter as either
   having AVG_DISTINCT selectivity; or the AVG_DISTINCT selectivity has
   been adjusted with cardinality (see below). This is necessary since
   when the join selectivity is calculated in the beginning we don't
   know the cardinality of the collection yet (after application of all
   table filters), and thus we don't use cardinality information in the
   beginning. Later on when this information is available, we in-effect
   recalculate this join selectivity
 - when calculating (1/distincts) as average join selectivity, add
   capability to take into consideration the cardinality (if available),
   by treating the cardinality as a "sample", and use the same formula
   as UPDATE STATISTICS in calculation of number of distincts based on
   a sample
 - mark join selectivity calculated through (1/distincts) with the new
   flag
 - during join enumeration, when "baseLevel" is available (which
   represents all single-term plans), use the cardinality information
   for a collection to re-recalculate join selectivity if it was
   marked with the new filter flag for AVG_DISTINCT. Mark the filter
   with a second Filter flag ADJ_DISTINCT afterwards.
 - add a new utility function to retrieve histogram from dictionary
   (existing common code in several places)
 - pass advisorValidate to join enumeration to mark validate phase
   of ADVISE

Change-Id: I967d6e5bb8e82cc8716118f487507b15bab9fba0
Reviewed-on: https://review.couchbase.org/c/query/+/225301
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Well-Formed: Restriction Checker