MB-65954 Add formalization of subquery expressions after expression parsing
The issue with this bug is with lack of certain identifier flags during
query planning. We have a query with a WITH clause, and when the query
gets prepared the query plan that gets shipped to another query node
only contains the main query block. The WITH clause subquery is shipped
as expression text. On the receiving node, the shipped query plan is
unmarshalled. When this is done, the entire subquery for the WITH
clause is simply parsed as an expression, then the subquery get planned.
When covering determination is done, for a keyspace that is being
joined, the covering process should have recognized that it is a
different keyspace thus "CoveredSkip" should be returned. However,
this depends on the keyspace identifier being marked with a KEYSPACE
flag, which is not present. This results in incorrect covering
determination.

The KEYSPACE flag on the identifier is set as part of the formalization
process. Although technically we don't need to formalize again since it
has gone through that when the statement is prepared on the first node
(i.e. we know the statement is "proper" and all keyspace prefixes have
been added), there are other processing done during formalization is
still needed, thus the fix is to add a formalization step during
parsing for a subquery expression, only when expression parsing is
being done, and only for uncorrelated subquery (for correlated subquery
we already go through a formalization process in order to check for and
mark proper correlation).

Change-Id: I46fd34e92ce740114e3ed94d3887452bb6c6c5e8
Reviewed-on: https://review.couchbase.org/c/query/+/225406
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-on: https://review.couchbase.org/c/query/+/225566