MB-65976 Derive filters for outer join on clause to enable secondary   ... index scans during join enumeration
For planning of a join query we derived IS NOT NULL predicate from
join predicate to facilitate index selection. This is currently only
done for inner joins. For an outer join, this is not an issue if
join enumeration is not in play, since when planner considers the
join we already moved the join predicate into regular filters such
that it can be used for index selection. For join enumeration, this
can be an issue since we construct single-keyspace plans at the
beginning of join enumeration, and thus we may need the derived
filter for index selection.

Prior to the fix for MB-62274, we used to derive predicate for outer
join as well (unintentionally), however with the fix for MB-62274
this is no longer performed. To allow more complete index selection
for join enumeration, we now reverse the original fix for MB-62274
and fix that issue in an alternative manner:
 - allow deriving IS NOT NULL predicate for outer table (reverse
   original fix for MB-62274)
 - when getting filters for joins, add logic to skip the derived
   IS NOT NULL filter in case of outer join
 - similarly for Lookup Join constructed as part of ANSI join
   consideration (join on meta().id), add similar logic to skip
   derived filter in case of outer join, when constructing an
   ON-clause filter for the Lookup Join
 - also added transfering of SubTimes after join enumeration

Change-Id: I554bcbbefd7960a99b5a70a7b6bae44b322271b6
Reviewed-on: https://review.couchbase.org/c/query/+/225474
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker